Fixed and improved toolset-based mangling for MSVS with LLVM toolset
LLVM toolset for MSVS compiles the code with clang-cl that produces executables
and libraries that are binary compatible with those compiled with original MSVC
compiler. In this case, it may be prefered that the target artefacts name
mangling to be the same as for the compatible MSVC toolset.

The new fixed and improved version of the utk_cmake_name_mangling_postfix()
function detects LLVM toolset and adds options for adjusting the generated name
mangling postfix. The option allows selecting from LLVM toolset name and MSVC
toolset name with the latter being the prefered option value.

For LLVM-vs2014 toolset two versions of MSVC toolset mangling are available:
vs140 and vc141. This is due to the fact that MCVC 14.0 and MSVC 14.1 are binary
compatible.

Signed-off-by: Innokentiy Alaytsev <alaitsev@gmail.com>