[6] Implemented a framework for dependency handling in *-target-config.cmake
The framework consists of two parts: a set of variables to control dependency
handling and the *-target-config.cmake.in template file.

The control variables are:

 * FIND_TARGET_DEPENDENCIES - the contents of the variable are used to find
   dependencies of the target.

 * TARGET_APPLY_DEPENDENCIES_TO_THE_CURRENT_DIRECTORY - the contents of the
   variable are used to apply the targets dependencies with calls to the
   include_directories() and link_libraries() functions.

 * TARGET_APPLY_IMPORTED_TARGET_DEPENDENCIES - the contents of the variable are
   used to apply dependencies to the imported target.

The FIND_TARGET_DEPENDENCIES variable is always used.

The TARGET_APPLY_DEPENDENCIES_TO_THE_CURRENT_DIRECTORY is used when subprojects
of the original project are built. This is useful for tests and examples:
instead of relying on the presence of the original target these projects may use
find_package() to find the target. The main advantage of it is that it allows
checking if the main target is properly exported and imported.

The TARGET_APPLY_IMPORTED_TARGET_DEPENDENCIES variable is used when the target
is actually imported after being found by the find_package() or
find_dependency() function.

The *-target-config.cmake.in implements application of the appropriate set of
dependencies.

Signed-off-by: Innokentiy Alaytsev <alaitsev@gmail.com>