Improved handling of multiple versions of the downloaded dependencies
In the previous version, problems could arise while trying to download a new
version of the dependency after some other version has already been
downloaded. Manual removal of the previous version could resolve the problem. To
minimise the need for user intervention a new mechanism for placing the
downloaded dependencies is introduced: each version of each dependency is put in
a separate directory with the name based on the download options used to
download this version of the dependency (MD5 hash).

Signed-off-by: Innokentiy Alaytsev <alaitsev@gmail.com>